refactor(activity-logs): break to files
